changing table

+++ database/migrations/2024_09_04_081807_add_three_columns_to_guards.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::table('guards', function (Blueprint $table) {
             $table->string('iban')->nullable();
             $table->string('bank')->nullable();
-            $table->float('salary')->nullable();
+            $table->decimal('salary', 6, 2)->nullable();
         });
     }
